apis/rag_memory/main.py
"""
Main entry point for RAG Memory Service API.
"""
import uvicorn
from api import app
from dotenv import load_dotenv
load_dotenv()

if __name__ == "__main__":
    uvicorn.run(
        "api:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )
