apis/rag_memory/requirements.txt
# Core dependencies
openai==2.8.1
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
pgvector==0.4.1
python-dotenv==1.2.1

# FastAPI and web server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data processing
numpy==2.2.6
networkx==3.4.2

# HTTP client
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0

# Type validation
pydantic==2.12.4
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# Utilities
tqdm==4.67.1
certifi==2025.11.12
idna==3.11
anyio>=3.7.1,<5.0.0
sniffio==1.3.1
distro==1.9.0
exceptiongroup==1.3.1
jiter==0.12.0

